configurando testes
