--- requirements.txt
@@ -1,8 +1,0 @@
-alembic>=1.16.2
-fastapi>=0.115.14
-httpx>=0.28.1
-jinja2>=3.1.6
-pydantic-settings>=2.10.1
-python-multipart>=0.0.20
-sqlmodel>=0.0.24
-uvicorn>=0.34.3
